+++ src/Skyscanner/JsonPath/InvalidJsonException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace JsonPath;
+
+/**
+ * Exception that is raised when a invalid value is given to the JsonObject
+ * constructor.
+ *
+ * @uses Exception
+ */
+class InvalidJsonException extends \Exception
+{
+}